Add files via upload
This is a college approval system in which a user should register first in order to access the application later he will be forwarded to login page .After logging in login page redirects the user to home page where he finds an option for applying for a college. After he fills a form the details are redirected to approver .Approver checks his eligibility for college and will be approving him accordingly.